poortext: split richtext's layout code into new package
The richtext package currently has two duties: laying out spans with
different styles, and providing an abstraction for pointer input
handling. The input handling is fairly opinionated and users may be
interested in implementing their own, while still reusing the existing
layout logic.

This package is a version of richtext that doesn't concern itself with
input handling. Instead, it allows users to pass in a function that gets
called for each span, letting users add their own input ops.

We've experimented with the following alternatives to providing a single
callback:

1. Storing a callback per span, in SpanStyle. This approach has the
downside of causing closures to escape to the heap, which results in
additional allocations for each span.

2. Storing pointer.InputOp and keyboard.InputOp in SpanStyle. This
doesn't compose well with the existing Gio ecosystem. For example,
widget.Click expects to be able to add input ops itself, via an Add
method. This wouldn't be possible if the user were required to provide
input ops upfront. It also limits users in what kinds of input ops they
can add.

3. Instead of providing a slice of spans and a callback, we experimented
with providing one span and a callback at a time. This doesn't work
because laying out spans with differing sizes is a two-pass process,
with the first pass finding line heights and the second pass actually
laying out spans.